refactor: Use explicit types for breadcrumb data
Replace raw `uint64_t` fields with `TimestampNs`, `ThreadId`, and
`FlowId` type aliases across the breadcrumb system. Update the
`BreadcrumbEnvelope` struct to use the `BreadcrumbPhase` and
`PayloadType` enums instead of `uint8_t`, which removes the need for
`static_cast` during assignments.

Establishing these explicit types clarifies the intended data
representation and enables compiler-level type checking for the crash
report breadcrumb structures.

+++ emulator/crashreport/tool/breadcrumbs/breadcrumbs_test.cc
@@ -55,8 +55,8 @@ TEST(BreadcrumbParserTest, ParsesContiguousBuffer) {
     envelope.timestamp_ns = 1000;
     envelope.thread_id = 123;
     envelope.flow_id = 456;
-    envelope.phase = static_cast<uint8_t>(BreadcrumbPhase::kFlowBegin);
-    envelope.payload_type = static_cast<uint8_t>(PayloadType::kGrpcProto);
+    envelope.phase = BreadcrumbPhase::kFlowBegin;
+    envelope.payload_type = PayloadType::kGrpcProto;
     envelope.payload_len = payload_len;
 
     uint32_t total_size = sizeof(BreadcrumbEnvelope) + payload_len;
